Fix exported types namespace
